producer/requirements.txt
requests==2.31.0
kafka-python==2.0.2
pyyaml==6.0.1
